feat(APIM-96): added GET `/yield-rates` endpoint (#50)
### Introduction
- New endpoint GET /yield-rates. It calculates UKEF exposure in months
for EW and BS Facilities
  Query parameters: searchDate (optional)

### Resolution
- GET /exposure-period uses TypeORM to get data from table
DWR_YIELD_RATE
- Old endpoint had query parameter searchDatetime, I changed it to
searchDate
- Old endpoint had field short_name, I renamed to shortName


### Miscellaneous
- Removed @ApiParam from all modules because Swagger module can generate
URL and Query parameter documentation from DTO class. We need URL and
Query parameters in DTO for validation, so we also keep it there for
documentation.
- Change MAXIMUM_TIMEZONE_LIMIT format to better match what DB response
looks, and this allows to use it for testing.
- Changed GET /exposure-period Query parameter type from date to string
date, it allows easier way to stop time input.