feat(APIM-162): API management instance (#78)
## Introduction
An API management gateway is a necessity in order to authenticate,
authorize, cache and throttle endpoints. APIM will also provide a façade
for the back-end endpoint hosting containers.

## Resolution
* Azure API management instance creation and configuration, APIM `sku`
a.k.a plan will be set respective to the targeted deployment
environment. `Developer` for development and staging, whereas a
pre-defined plan for `production`.
* APIM and CA configuration.
* Enhanced VNET peering creation.
* Naming convention update.
* Base docker image updated to `19.9`.
* Updated `@ApiProperty` from `examples` array to `example` string, due
to OpenAPI 3.0 validation error.
* Added `jsonDocumentUrl` for custom direct OpenAPI JSON export.

## Miscellaneous
* Spell fixes
* Deployment script: added staging
* Added GHA files to deployment paths for both `infrastructure` and
`deployment`.
* Various dependencies updates.

## Future
* Import APIM policies as IaC.

---------

Co-authored-by: Abhi Markan <abhi.markan@ukexportfinance.gov.uk>