feat(FN-3163): automate new DTFS customer creation in Salesforce (D&B service only) (#1130)
## Introduction :pencil2:
As part of the CCM project, if a customer does exist in SF, we are
automatically creating it rather than relying on the manual process
currently undertaken by Business Support Group. DTFS will send requests
to a new customer/salesforce endpoint for MDM to create a customer in
Salesforce. Technical documentation of this process is
[here](https://ukef.sharepoint.com/:w:/r/sites/pmo-PRO211/Shared%20Documents/1.%20TO%20USE%20New%20file%20plan/2.%20Tranche%202/03.%20Customer%20and%20Credit%20Migration/Documents/DTFS%20%26%20Salesforce%20Integration/DTFS%20-%20SF%20Integration%20Technical%20Details.docx?d=w75ec8ca455ec492db2a912cc20cea7e7&csf=1&web=1&e=rhuzKr)

The corresponding feature-flagged DTFS pull request is
[here](https://github.com/UK-Export-Finance/dtfs2/pull/4021)

## Resolution :heavy_check_mark:
This PR adds the Dun and Bradstreet service, and an endpoint to allow us
to GET a company from Dun and Bradstreet by their registration number
(in our case Companies House Number). This will then be used by [this
PR](https://github.com/UK-Export-Finance/mdm-api/pull/1057) as part of
the customer creation process.

## Miscellaneous :heavy_plus_sign:
To be merged before the remainder of the customer creation endpoint.

Also adds a user facing D&B endpoint, viewable in the Swagger docs:

![image](https://github.com/user-attachments/assets/09829422-4e50-4a81-807e-60651cd099a0)

---------

Co-authored-by: Nat Dean-Lewis <ndlewis@ukexportfinance.gov.uk>