feat(FN-3163): add get or create customer endpoint (#1132)
## Introduction :pencil2:
As part of the CCM project, if a customer doesn't exist in SF, we are
automatically creating it rather than relying on the manual process
currently undertaken by Business Support Group. DTFS will send requests
to a new customer/salesforce endpoint for MDM to create a customer in
Salesforce. Technical documentation of this process is
[here](https://ukef.sharepoint.com/:w:/r/sites/pmo-PRO211/Shared%20Documents/1.%20TO%20USE%20New%20file%20plan/2.%20Tranche%202/03.%20Customer%20and%20Credit%20Migration/Documents/DTFS%20%26%20Salesforce%20Integration/DTFS%20-%20SF%20Integration%20Technical%20Details.docx?d=w75ec8ca455ec492db2a912cc20cea7e7&csf=1&web=1&e=rhuzKr)

## Resolution :heavy_check_mark:
This PR adds a "get or create" salesforce customer endpoint that does a
few things:
Checks for an existing customer via Informatica. If none exists (or we
return a legacy one):
Generates a URN from NUMGEN (if there isn't already one returned from a
legacy record by Informatica, in which case we just use that).
Calls the Dun and Bradstreet service to get a customer's DUNS number.
POSTs these along with the customer name and Companies House Number to
Salesforce.
Returns the customer, including the URN to DTFS for deal submission.

## Miscellaneous :heavy_plus_sign:
A tweak to the NUMGEN regex to allow customer URN creation, and a typo
fix.

---------

Co-authored-by: Nat Dean-Lewis <ndlewis@ukexportfinance.gov.uk>