feat(FN-3163): remove local ssl workaround (#1188)
## Introduction :pencil2:
On local builds, the Salesforce domains needed to create a customer are
not whitelisted on the UKEF VPN, so rejectUnauthorized must be set as
false. This doesn't affect higher environments.

## Resolution :heavy_check_mark:
Remove this setting so local builds will fail, but will replicate the
code running on higher environments more closely.

## Miscellaneous :heavy_plus_sign:
Remove the same commented-out behaviour from Informatica, this domain IS
whitelisted so works on local builds regardless.

Co-authored-by: Nat Dean-Lewis <ndlewis@ukexportfinance.gov.uk>