chore(APIM-612): companies house - isactive field (#1182)
## Introduction :pencil2:
In order for Companies House to be consumed by other products/services,
the integration should return an `isActive` field.

This will ensure that any product/service that needs to check if a
company is active, will not need to perform a check on the
`company_status` field and instead, can simply consume a `isActive`
field.

## Resolution :heavy_check_mark:
- Update the service to return a `isActive` field.
- Update types and `GetCompanyGenerator`.