feat(FN-3163): allow non gef deals (#1201)
## Introduction :pencil2:
From QA testing it was recognised that it would be good to allow non-GEF
deals to automatically create customers in Salesforce (as well as GEF
deals which already have this behaviour implemented). The only blocker
to this is that Probability of Default needs to be optional as non-GEF
exporters don't have a Probability of Default entered.

## Resolution :heavy_check_mark:
Make Probability of Default optional.

---------

Co-authored-by: Nat Dean-Lewis <ndlewis@ukexportfinance.gov.uk>